carbonless_backend/requirements.txt
django>=5.0,<7.0
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.3,<5.0
django-ratelimit>=4.1,<5.0
gunicorn>=21.0,<23.0
whitenoise>=6.5,<7.0
dj-database-url>=2.1,<3.0
psycopg2-binary>=2.9,<3.0
reportlab>=4.0,<5.0
openpyxl>=3.1,<4.0
